Further broke up the previously created ActivityTest
